backend/requirements.txt
cerebras-cloud-sdk
python-dotenv
openai

anyio
fastapi
h11
importlib-metadata
pydantic
sniffio
starlette
typing-extensions
uvicorn
python-multipart
requests==2.32.3

numpy
pandas

alembic==1.15.2
temporalio==1.4.0
openai==0.28
langchain==0.2.1
langchain-community==0.2.1
langchain-core==0.2.1
langchain-text-splitters==0.2.0
langcodes==3.4.0
langsmith==0.1.63
language_data==1.2.0
stripe==11.6.0
langchain_community
